Partially move `test_conftest` to `test__deprecation`
