+++ pvlib/tests/test__deprecation.py
@@ -5,8 +5,54 @@ Test the _deprecation module.
 import pytest
 
 from pvlib import _deprecation
+from .conftest import fail_on_pvlib_version
 
 import warnings
+
+
+@pytest.mark.xfail(strict=True,
+                   reason='fail_on_pvlib_version should cause test to fail')
+@fail_on_pvlib_version('0.0')
+def test_fail_on_pvlib_version():
+    pass
+
+
+@fail_on_pvlib_version('100000.0')
+def test_fail_on_pvlib_version_pass():
+    pass
+
+
+@pytest.mark.xfail(strict=True, reason='ensure that the test is called')
+@fail_on_pvlib_version('100000.0')
+def test_fail_on_pvlib_version_fail_in_test():
+    raise Exception
+
+
+# set up to test using fixtures with function decorated with
+# conftest.fail_on_pvlib_version
+@pytest.fixture
+def some_data():
+    return "some data"
+
+
+def alt_func(*args):
+    return args
+
+
+@pytest.fixture
+def deprec_func():
+    return _deprecation.deprecated(
+        "350.8", alternative="alt_func", name="deprec_func", removal="350.9"
+    )(alt_func)
+
+
+@fail_on_pvlib_version('350.9')
+def test_use_fixture_with_decorator(some_data, deprec_func):
+    # test that the correct data is returned by the some_data fixture
+    assert some_data == "some data"
+    with pytest.warns(_deprecation.pvlibDeprecationWarning):
+        # test for custom deprecation warning provided by pvlib
+        deprec_func(some_data)
 
 
 @pytest.fixture
